requirements.txt
# anyio==3.6.2
# appdirs==1.4.4
# bcrypt==3.2.0
# beautifulsoup4==4.12.2
# blinker==1.7.0
# certifi==2020.6.20
# chardet==4.0.0
# click==8.1.7
# colorama==0.4.4
# cryptography==3.4.8
# distlib==0.3.4
# docutils==0.17.1
# entrypoints==0.4
# fasteners==0.14.1
# filelock==3.6.0
Flask==3.0.2
# future==0.18.2
# h11==0.14.0
# h2==4.1.0
# hpack==4.0.0
# httpcore==0.16.3
# httpx==0.23.3
# hyperframe==6.0.1
# idna==3.3
# importlib-metadata==4.6.4
# itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
# keyring==23.5.0

python-telegram-bot==20.1
requests==2.25.1
telebot==0.0.5

